% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleBinary.R
\name{bi_es_AR}
\alias{bi_es_AR}
\title{Single Binary Effect Size - Alternative Ratio}
\usage{
bi_es_AR(data, field)
}
\arguments{
\item{data}{A dataframe}

\item{field}{The field within the dataframe to use as string}
}
\value{
results A dataframe with the following information

AR1 The Alternative Ration for the first category

AR2 The Alternative Ration for the second category
}
\description{
This function performs calculates the Alternative Ratio (a.k.a. Relative Risk).
}
\details{
JonB on CrossValidated suggests to use Relative Risks as effect size measure
for a binomial test, which the NCSS calls Alternative Ratio (n.d.).

The Alternative Ratio can be calculated by:
\deqn{AR = \frac{p}{\pi}}
It is simply the actual proportion (:math:\code{p}) divided by the expected proportion (\eqn{\pi}).
If the two are equal, the result will be 1.
For the interpretation it is therefor easier to look at the difference with 1.
It then informs you how much more or less likely the category is than was expected.

For more details on Alternative Ratio with R see: https://youtu.be/V5tug-2Zqjc
}
\references{
JonB. (2015, October 14). Effect size of a binomial test and its relation to other measures of effect size. Retrieved from https://stats.stackexchange.com/q/176856

NCSS. (n.d.). Tests for two proportions using effect size. In PASS Sample Size Software (pp. 199-1-199–10). Retrieved from https://ncss-wpengine.netdna-ssl.com/wp-content/themes/ncss/pdf/Procedures/PASS/Tests_for_Two_Proportions_using_Effect_Size.pdf
}
\author{
P. Stikker
Please visit: https://PeterStatistics.com
YouTube channel: https://www.youtube.com/stikpet
}
