% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleBinary.R
\name{bi_es_coheng}
\alias{bi_es_coheng}
\title{Single Binary Effect Size - Cohen's g}
\usage{
bi_es_coheng(data, field)
}
\arguments{
\item{data}{A dataframe}

\item{field}{The field within the dataframe to use as string}
}
\value{
list With a qualification and the effect size measure
}
\description{
This function performs calculates Cohen's g.
}
\details{
Rosnow and Rosenthal (2003) mention as effect size for binary (they call it dichotomous)
data Cohen's g and Cohen's h.

Cohen's g can be calculated by:
\deqn{g = p_i - 0.50}
The \eqn{p_i} is simply the proportion of category i.

The maximum value for Cohen's g would be 0.5, which would occur
if the sample proportion for one of the two categories is either 0 (0\%) or 1 (100\%).

Cohen (1988, p. 147) actually gives thresholds for 'small', 'medium', and 'large' effect sizes
for this g at .05, .15, and .25 resp. (Cohen, 1988, pp. 147-149). I would interpret this as:\tabular{ll}{
   Cohen g \tab qualification \cr
   0 < .05 \tab Negligible \cr
   0.10 < 0.15 \tab Small \cr
   0.15 < 0.25 \tab Medium \cr
   0.25 or more \tab Large \cr
}


For more info on Cohen's g with R: https://youtu.be/25CQR5OQTK0
}
\references{
Cohen, J. (1988, p. 147). \emph{Statistical power analysis for the behavioral sciences} (2nd ed). Hillsdale, N.J: L. Erlbaum Associates.

Rosnow, R. L., & Rosenthal, R. (2003). Effect sizes for experimenting psychologists. \emph{Canadian Journal of Experimental Psychology/Revue Canadienne de Psychologie Expérimentale, 57}(3), 221–237. doi:10.1037/h0087427
}
\author{
P. Stikker
Please visit: https://PeterStatistics.com
YouTube channel: https://www.youtube.com/stikpet
}
