% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleBinary.R
\name{bi_es_cohenH2}
\alias{bi_es_cohenH2}
\title{Single Binary Effect Size - Cohen's H_2}
\usage{
bi_es_cohenH2(data, field)
}
\arguments{
\item{data}{A dataframe}

\item{field}{The field within the dataframe to use as string}
}
\value{
list With a qualification and the effect size measure
}
\description{
This function performs calculates Cohen's H_2.
}
\details{
Rosnow and Rosenthal (2003) mention as effect size for binary (they call it dichotomous)
data Cohen's g and Cohen's h. Cohen's h is also the effect size used for two proportions in the NCSS software (n.d.).

Cohen's h from one sample can be calculated by (Cohen, 1988, p. 202):
\deqn{h_2 = \phi_1 - \phi_c}
with \eqn{\phi_i = 2\times arcsin\left(\sqrt{p_i}\right)}

The \eqn{p_i} is simply the proportion of category i and
\eqn{p_c} the expected proportion in the population for that category.

Cohen actually gives thresholds for 'small', 'medium', and 'large' effect sizes for h at .20, .50, and .80 resp. (Cohen, 1988, pp. 184-185).
I would interpret this as:\tabular{ll}{
   Cohen h \tab qualification \cr
   0 < .20 \tab Negligible \cr
   0.20 < 0.50 \tab Small \cr
   0.50 < 0.80 \tab Medium \cr
   0.80 or more \tab Large \cr
}


Note that this table is for Cohen's h for paired data. He gives a conversion that can be used from the one-sample:
\deqn{h = h_2 \times \sqrt{2}}

For more info on Cohen's h2 with R: https://youtu.be/e6JanIx0xEY
}
\references{
Cohen, J. (1988). \emph{Statistical power analysis for the behavioral sciences} (2nd ed). Hillsdale, N.J: L. Erlbaum Associates.

Rosnow, R. L., & Rosenthal, R. (2003). Effect sizes for experimenting psychologists. \emph{Canadian Journal of Experimental Psychology/Revue Canadienne de Psychologie Expérimentale, 57}(3), 221–237. doi:10.1037/h0087427

NCSS. (n.d.). Tests for two proportions using effect size. In PASS Sample Size Software (pp. 199-1-199–10). Retrieved from https://ncss-wpengine.netdna-ssl.com/wp-content/themes/ncss/pdf/Procedures/PASS/Tests_for_Two_Proportions_using_Effect_Size.pdf
}
\author{
P. Stikker
Please visit: https://PeterStatistics.com
YouTube channel: https://www.youtube.com/stikpet
}
