% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleBinary.R
\name{bi_ts_chi2Gof_G}
\alias{bi_ts_chi2Gof_G}
\title{Single Binary Test - G Goodness of Fit test}
\usage{
bi_ts_chi2Gof_G(data, field, corr = "none")
}
\arguments{
\item{data}{A dataframe}

\item{field}{The field within the dataframe to use as string}

\item{corr}{Correction to use, either 'none' (default), 'yates' for Yates, 'pearson' for Pearson or 'williams' for Williams}
}
\value{
result A dataframe with the following:

chi2Val The chi-square test statistic

df The degrees of freedom

n The sample size

pVal The p-value (significance)
}
\description{
This function performs a G Goodness of Fit test (a.k.a. likelihood ratio GoF).
It assumes the probability of success in the population to be 0.5
}
\details{
The test uses:
\deqn{G = 2\times\sum_{i=1}^{k}\left(O_i\times ln\left(\frac{F_i}{E_i}\right)\right)}
Where \eqn{k} is the number of categories (i.e. 2), \eqn{F_i} is the observed frequency of category i,
and :math:\code{E_i} the expected frequency (i.e. n/2)

For the Yates continuity correction the formula is adjusted to:
\deqn{G_{Yates} = 2\times \sum_{i=1}^k\left(F_i^{'}\times ln\left(\frac{F_i^{'}}{E_i}\right)\right)}
with:
\deqn{F_i^{'}=\begin{cases} F_i-0.5 & \text{ if } F_i> E_i \\ F_i+0.5 & \text{ if } F_i<E_i \\ F_i & \text{ if } F_i=E_i\end{cases}}

For the E. Pearson correction:
\deqn{G_{\text{E.Pearson}} = G\times\frac{n-1}{n}}

For the Williams correction:
\deqn{G_{\text{Williams}} = \frac{G}{q}}
with:
\deqn{q = 1+\frac{k^2-1}{6\times n\times df}}

For more info on this test with R: https://youtu.be/27n2ko-S5U0
}
\author{
Author: P. Stikker
Please visit: https://PeterStatistics.com
YouTube channel: https://www.youtube.com/stikpet
}
