% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleBinary.R
\name{bi_ts_chi2Gof_Pearson}
\alias{bi_ts_chi2Gof_Pearson}
\title{Single Binary Test - Pearson Chi-Square Goodness of Fit test}
\usage{
bi_ts_chi2Gof_Pearson(data, field, corr = "none")
}
\arguments{
\item{data}{A dataframe}

\item{field}{The field within the dataframe to use as string}

\item{corr}{Correction to use, either 'none' (default), 'yates' for Yates, 'pearson' for Pearson or 'williams' for Williams}
}
\value{
list With the chi-square value, degrees of freedom, sample size, and the two-sided significance (p-value)
}
\description{
This function performs a Pearson Chi-Square Goodness of Fit test.
It assumes the probability of success in the population to be 0.5
}
\author{
P. Stikker
Please visit: https://PeterStatistics.com
YouTube channel: https://www.youtube.com/stikpet
}
