% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleBinary.R
\name{bi_ts_score}
\alias{bi_ts_score}
\title{Single Binary Test - One-Sample Score test}
\usage{
bi_ts_score(data, field, corr = FALSE)
}
\arguments{
\item{data}{A dataframe}

\item{field}{The field within the dataframe to use as string}

\item{corr}{A boolean TRUE or FALSE to indicate the use of a Yates correction}
}
\value{
result A dataframe with the following:

zVal The z-statistic used

pVal The p-value of the test (two sided)
}
\description{
This function performs a one-sample score test.
It assumes the probability of success in the population to be 0.5
}
\details{
The test uses:
\deqn{z = \frac{k - \mu}{\sigma}}
With \eqn{\mu = n\times p_0}, and \eqn{\sigma = \sqrt{\mu \times \left( 1 - p_0\right)}}
Where \eqn{k} is the number of successes, \eqn{n} is the total sample size and \eqn{p_0} is the proportion according to the null hypothesis (set to 0.5)
For the Yates continuity correction the formula is adjusted to:
\deqn{z_{Yates} = \frac{\left|x - \mu\right| - 0.5}{\sigma}}

For more info on this test with R: https://youtu.be/27n2ko-S5U0
}
\author{
P. Stikker
Please visit: https://PeterStatistics.com
YouTube channel: https://www.youtube.com/stikpet
}
