% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleBinary.R
\name{bi_ts_wald}
\alias{bi_ts_wald}
\title{Single Binary Test - One-Sample Wald test}
\usage{
bi_ts_wald(data, field, corr = FALSE)
}
\arguments{
\item{data}{A dataframe}

\item{field}{The field within the dataframe to use as string}

\item{corr}{A boolean TRUE or FALSE to indicate the use of a Yates correction}
}
\value{
result A dataframe with the following:

zVal The z-statistic used

pVal The p-value of the test (two sided)
}
\description{
This function performs a one-sample Wald test.
It assumes the probability of success in the population to be 0.5
}
\details{
The test uses:
\deqn{z = \frac{k - \mu}{s}}
With \eqn{\mu = n\times p_0}, and \eqn{s = \sqrt{k\times\left(1 - \frac{k}{n}\right)}}
Where \eqn{k} is the number of successes, \eqn{n} is the total sample size
and \eqn{p_0} is the proportion according to the null hypothesis (set to 0.5)
For the Yates continuity correction the formula is adjusted to:
\deqn{z = \frac{\left|k - \mu\right| - 0.5}{s}}

For more info on this test with R: https://youtu.be/27n2ko-S5U0
}
\author{
P. Stikker
Please visit: https://PeterStatistics.com
YouTube channel: https://www.youtube.com/stikpet
}
