% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleNominal.R
\name{bin_decider}
\alias{bin_decider}
\title{Single Nominal - Post-Hoc - Binary Settings}
\usage{
bin_decider(n, nThres = 1000, prefTest = "g", corr = "yates", es = "cohenH2")
}
\arguments{
\item{n}{Sample size used in test}

\item{nThres}{An integer specifying a threshold up to which to use the exact binomial test}

\item{prefTest}{Indicator for the preferred approximation, either 'g' (default) or 'pearson'}

\item{corr}{Indicator for which correction to apply either 'yates' (default), or 'pearson' for Pearson or 'williams' for Williams}

\item{es}{Indicator for effect size measure to use in post-hoc test, either 'cohonH2' (default), 'coheng', or 'ar'}
}
\value{
results a data frame with:

test The name of the test to use

corr The correction to use in the test

es The effect size measure to use
}
\description{
This is a helper function to decide on which post-hoc test to use.
}
