% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleBinary.R
\name{frequencyTable}
\alias{frequencyTable}
\title{Frequency Table
Generates a frequency table with the same info as SPSS would}
\usage{
frequencyTable(data, field)
}
\arguments{
\item{data}{A dataframe}

\item{field}{The field within the dataframe to use as string}
}
\value{
table A table with frequencies, percentage, valid percentage and cumulative percentage
}
\description{
Frequency Table
Generates a frequency table with the same info as SPSS would
}
\details{
For more info on this test with R: https://youtu.be/VjJCzeRa49w
}
\author{
P. Stikker
Please visit: https://PeterStatistics.com
YouTube channel: https://www.youtube.com/stikpet
}
