% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleNominal.R
\name{nom_ce_mode}
\alias{nom_ce_mode}
\title{Single Nominal - Central Tendancy - Mode}
\usage{
nom_ce_mode(data, field)
}
\arguments{
\item{data}{Dataframe to be used}

\item{field}{Specific field name (as string) from data frame}
}
\value{
values The category or categories that are the mode
}
\description{
Function to determine the mode for a nominal variable
}
\details{
This function will return all modes if the data is multimodal, or no mode if all categories have equal count
}
