% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleNominal.R
\name{nom_di_vr}
\alias{nom_di_vr}
\title{Single Nominal - Dispersion - Variation Ratio}
\usage{
nom_di_vr(data, field)
}
\arguments{
\item{data}{Dataframe to be used}

\item{field}{Specific field name (as string) from data frame}
}
\value{
values The Variation Ratio
}
\description{
Simple function to determine the variation ratio of a nominal variable
}
\details{
The Variation Ratio (VR) (Freeman, 1965) is simply the proportion that does not belong to the modal category (Zedeck, 2014, p.406).
}
\references{
Freeman, L. C. (1965). \emph{Elementary applied statistics: for students in behavioral science}. Wiley.

Zedeck, S. (Ed.). (2014). \emph{APA dictionary of statistics and research methods}. American Psychological Association.
}
