% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleNominal.R
\name{nom_es_cramerv}
\alias{nom_es_cramerv}
\title{Single Nominal - Effect Size - Cramer's V}
\usage{
nom_es_cramerv(data, field, test = "pearson", corr = "none", bergsma = FALSE)
}
\arguments{
\item{data}{Dataframe to be used}

\item{field}{Specific field name (as string) from data frame}

\item{test}{Which test to use, either 'pearson' (default) or 'g'}

\item{corr}{Correction to use in test, either 'none' (default), 'yates', 'pearson', or 'williams'}

\item{bergsma}{Boolean to indicate to use the Bergsma correction on Cramer's V, FALSE (default) or TRUE}
}
\value{
results A dataframe with:

qual A qualification of the effect size

es The effect size
}
\description{
This function will determine Cramer's V for a Goodness-of-Fit test
}
