% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleNominal.R
\name{nom_ph_test}
\alias{nom_ph_test}
\title{Single Nominal - Post-Hoc}
\usage{
nom_ph_test(data, field, binDecider = c(1000, "g", "yates", "cohenH2"))
}
\arguments{
\item{data}{Dataframe to be used}

\item{field}{Specific field name (as string) from data frame}

\item{binDecider}{A vector with the settings for the binDecider, default is c(1000, 'g', 'yates', 'cohenH2')}
}
\value{
results A dataframe with all the information
}
\description{
This function will perform a post-hoc analysis on a nominal variable
}
