% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleNominal.R
\name{nom_ts_pearson}
\alias{nom_ts_pearson}
\title{Single Nominal - Test - Pearson chi-square goodness-of-fit}
\usage{
nom_ts_pearson(data, field, corr = "none")
}
\arguments{
\item{data}{Dataframe to be used}

\item{field}{Specific field name (as string) from data frame}

\item{corr}{Specific correction to be used, 'none' (default), 'yates', 'pearson', or 'williams'}
}
\value{
results A dataframe with the following values

chi2Val The chi-square test statistic

pVal The p-value of the test (significance)

df The degrees of freedom

n The sample size

nExpBelow5 The number of cells with an expected count less than 5

pExpBelow5 The proportion of cells with an expected count less than 5

minExp The minimum expected count
}
\description{
This function will perform a Pearson chi-square goodness-of-fit (GoF) test on a nominal variable.
}
\details{
This function will assume that the expected counts are simply the sample size
divided by the total number of categories (i.e. uniform distributed)

The Pearson chi-square test formula is:
\deqn{ \chi^{2}=\sum_{i=1}^{k}\frac{\left(F_{i}-E_{i}\right)^{2}}{E_{i}}}
#' With E. Pearson correction:
\deqn{\chi_{E Pearson}^2 = \frac{n - 1}{n}\times \chi^2}
With Williams correction:
\deqn{\chi_{Williams}^2 = \frac{\chi^2}{q}}
Where:
\deqn{q = 1 + \frac{k+1}{6\times n}}
In the formulas \eqn{F_i} are the observed frequencies, \eqn{E_i} the expected frequencies,
\eqn{k} the number of categories, and \eqn{n} the total sample size.

Note that a Yates correction is only applicable with df = 1, i.e. a binary variable. So it
is not incorporated here.
}
