% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleNominal.R
\name{nom_vi_bar}
\alias{nom_vi_bar}
\title{Single Nominal - Visualisation - Bar Chart}
\usage{
nom_vi_bar(data, field, height = "count", missing = FALSE)
}
\arguments{
\item{data}{Dataframe to be used}

\item{field}{Specific field name (as string) to draw chart from}

\item{height}{Determines if the height of the bars are counts ('count', default) or percentages ('percent')}

\item{missing}{Boolean to indicate to include the missing values (default is FALSE)}
}
\value{
bar chart
}
\description{
Simple function to create a bar-chart.
}
\details{
A bar-chart is defined as “a graph in which bars of varying height with spaces between them are used
to display data for variables defined by qualities or categories” (Zedeck, 2014, p. 20)
}
\references{
Zedeck, S. (Ed.). (2014). \emph{APA dictionary of statistics and research methods}. American Psychological Association.
}
