% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleNominal.R
\name{nom_vi_cleveland}
\alias{nom_vi_cleveland}
\title{Single Nominal - Visualisation - Cleveland Dot plot}
\usage{
nom_vi_cleveland(
  data,
  field,
  missing = FALSE,
  dotScale = 1,
  options = c(19, 4, 2)
)
}
\arguments{
\item{data}{Dataframe to be used}

\item{field}{Specific field name (as string) to draw chart from}

\item{missing}{Boolean to indicate to include the missing values (default is FALSE)}

\item{dotScale}{One dot represents how many data points, default is 1}

\item{options}{A vector with settings for pch (dot style), col (color), and pt.cex (size)}
}
\value{
a Cleveland dot plot
}
\description{
Simple function to create a Cleveland dot plot.
}
\details{
A Cleveland dot plot (Cleveland & McGill, 1987) is a bar chart where instead of bars a dot is placed at
the center of the top of the bar (and then the bars removed). It is a dot plot only showing the top dot.
}
\references{
Cleveland, W. S., & McGill, R. (1984). Graphical perception: Theory, experimentation, and application to the development of graphical methods.
\emph{Journal of the American Statistical Association, 79}(387), 531–554. doi:10.2307/2288400
}
