% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleNominal.R
\name{nom_vi_dot}
\alias{nom_vi_dot}
\title{Single Nominal - Visualisation - Dot plot}
\usage{
nom_vi_dot(
  data,
  field,
  missing = FALSE,
  dotScale = 1,
  options = c(19, 4, 2, 0.8)
)
}
\arguments{
\item{data}{Dataframe to be used}

\item{field}{Specific field name (as string) to draw chart from}

\item{missing}{Boolean to indicate to include the missing values (default is FALSE)}

\item{dotScale}{One dot represents how many data points, default is 1}

\item{options}{A vector with settings for pch (dot style), col (color), cex (size) and offset (space between each dot)}
}
\value{
dot plot
}
\description{
Simple function to create a dot plot.
}
\details{
According to the Oxford Dictionary of Statistics a dot plot can be defined as
"an alternative to a bar chart or line graph when there are very few data values.
Each value is recorded as a dot, so that the frequencies for each value can easily be counted" (Upton & Cook, 2014, p. 129).

You can let one dot represent multiple datapoints by setting the dotScale parameter. Also the size, color and offset can be set.
}
\references{
Zedeck, S. (Ed.). (2014). \emph{APA dictionary of statistics and research methods}. American Psychological Association.
}
