% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleNominal.R
\name{nom_vi_pareto}
\alias{nom_vi_pareto}
\title{Single Nominal - Visualisation - Pareto chart}
\usage{
nom_vi_pareto(data, field)
}
\arguments{
\item{data}{Dataframe to be used}

\item{field}{Specific field name (as string) to draw chart from}
}
\value{
a Pareto chart
}
\description{
Simple function to draw a Pareto chart
}
\details{
Unfortunately, there is no general agreed upon definition of a Pareto diagram. The most general description I’ve found was
by Kemp and Kemp (2004) who mention it is a name for a bar chart if the order of the bars have no meaning (i.e. for a nominal variable),
and they only mention that often the bars are then placed in decreasing order.

According to some authors a Pareto diagram is any diagram with the bars in order of size (Joiner, 1995; WhatIs.com, n.d.),
while others suggest that a line representing the cumulative relative frequencies should also be included (Weisstein, 2002).

Upton and Cook (2014) also add that the bars should not have any gaps, but many other authors ignore this.

I will use the following definition: a bar chart where the bars are placed in descending order of frequency.
Usually an ogive is added in the chart as well.

An ogive (oh-jive) is: "the graphs of cumulative frequencies" (Kenney, 1939).
}
\references{
Joiner. (1995). \emph{Pareto charts: Plain & simple}. Joiner Associates.

Kemp, S. M., & Kemp, S. (2004). \emph{Business statistics demystified}. McGraw-Hill.

Kenney, J. F. (1939). \emph{Mathematics of statistics; Part one}. Chapman & Hall.

Upton, G. J. G., & Cook, I. (2014). \emph{Dictionary of statistics} (3rd ed.). Oxford University Press.

Weisstein, E. W. (n.d.). Pareto Plot. Retrieved May 25, 2014, from http://mathworld.wolfram.com/ParetoPlot.html

WhatIs.com. (n.d.). What is Pareto chart (Pareto distribution diagram)? - Definition from WhatIs.com.
Retrieved April 20, 2014, from http://whatis.techtarget.com/definition/Pareto-chart-Pareto-distribution-diagram
}
