% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleNominal.R
\name{nom_vi_pie}
\alias{nom_vi_pie}
\title{Single Nominal - Visualisation - Pie Chart}
\usage{
nom_vi_pie(data, field, missing = FALSE)
}
\arguments{
\item{data}{Dataframe to be used}

\item{field}{Specific field name (as string) to draw chart from}

\item{missing}{Boolean to indicate to include the missing values (default is FALSE)}
}
\value{
a pie chart
}
\description{
Simple function to create a Pie Chart.
}
\details{
Most definitions of a pie-chart describe the shape. For example one definition is given as
“a graphic display in which a circle is cut into wedges with the area of each wedge being proportional to
the percentage of cases in the category represented by that wedge” (Zedeck, 2014, p. 260).
}
\references{
Zedeck, S. (Ed.). (2014). \emph{APA dictionary of statistics and research methods}. American Psychological Association.
}
