% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleOrdinal.R
\name{ord_ce_median}
\alias{ord_ce_median}
\title{Single Ordinal - Central Tendency - Median}
\usage{
ord_ce_median(data, field, order = "none")
}
\arguments{
\item{data}{A dataframe to be used}

\item{field}{Specific field in the dataframe as string}

\item{order}{A vector with the labels in order, or set to 'none'}
}
\description{
This function will return the median, defined as “the middle value in a distribution,
below and above which lie values with equal total frequencies or probabilities” (Porkess, 1991, p. 134).
}
\references{
Porkess, R. (1991). \emph{The HarperCollins dictionary of statistics}. HarperPerennial.
}
