% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleOrdinal.R
\name{ord_di_consensus}
\alias{ord_di_consensus}
\title{Single Ordinal - Dispersion - Consensus
.
Determines a measure of dispersion called consensus (Tastle & Wierman, 2007; Tastle, Wierman, & Rex Dumdum, 2005).
This measure ranges from 0 to 1. A zero would indicate a complete lack of consensus,
the number of people that tend towards one end of the ordinal variable (i.e. fully disagree) is then the same as
the number of people who tend to the other end (i.e. fully agree),
while consensus of one would indicate all respondents gave the same answer.}
\usage{
ord_di_consensus(data, field, order = "none")
}
\arguments{
\item{data}{A dataframe to be used}

\item{field}{Specific field in the dataframe as string}

\item{order}{A vector with the labels in order, or set to 'none'}
}
\value{
Cns The consensus measure
}
\description{
Single Ordinal - Dispersion - Consensus
.
Determines a measure of dispersion called consensus (Tastle & Wierman, 2007; Tastle, Wierman, & Rex Dumdum, 2005).
This measure ranges from 0 to 1. A zero would indicate a complete lack of consensus,
the number of people that tend towards one end of the ordinal variable (i.e. fully disagree) is then the same as
the number of people who tend to the other end (i.e. fully agree),
while consensus of one would indicate all respondents gave the same answer.
}
\details{
The formula for the Consensus is:
\deqn{Cns = 1 + \sum_{i=1}^{k}p_i \log_2\left(1-\frac{\left|X_i-\mu_X\right|}{d_X}\left)}
With:
\deqn{\mu_X = \frac{\sum_{i=1}^k X_i\times F_i}{n}}
\deqn{d_X = \text{max}\left(X_i\right)-\text{min}\left(X_i\right)}
Where \eqn{X_i} is the rank of category i
}
\references{
Tastle, W. J., & Wierman, M. J. (2007). Consensus and dissention: A measure of ordinal dispersion.
\emph{International Journal of Approximate Reasoning, 45}(3), 531–545. doi:10.1016/j.ijar.2006.06.024

Tastle, W. J., Wierman, M. J., & Rex Dumdum, U. (2005). Ranking ordinal scales using the consensus measure.
\emph{Issues in Information Systems, 6}(2), 96–102.
}
