% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleOrdinal.R
\name{ord_es_rosenthal}
\alias{ord_es_rosenthal}
\title{Single Ordinal - Effect Size - Rosenthal Correlation
.
Determines the Rosenthal Correlation}
\usage{
ord_es_rosenthal(data, field, order = "none", hypMed = "none", correct = FALSE)
}
\arguments{
\item{data}{A dataframe to be used}

\item{field}{Specific field in the dataframe as string}

\item{order}{A vector with the labels in order, or set to 'none'}

\item{hypMed}{The median according to the null hypotheses. If set to none the median of the categories will be used}

\item{correct}{Boolean to indicate the use of a Yates continuity correction}
}
\value{
results A dataframe with:

es The correlation coefficient

qual A qualification of the effect size
}
\description{
Single Ordinal - Effect Size - Rosenthal Correlation
.
Determines the Rosenthal Correlation
}
\details{
This effect size is suggested by various authors (Fritz et al., 2012, p. 12; Mangiafico, 2016; Simone, 2017; Tomczak, M., & Tomczak, E., 2014, p. 23; ).
The formula can be found in Rosenthal book (1991, p. 19) , so I will refer to as the Rosenthal correlation coefficient
(as to differentiate it with other correlation coefficients). Probably the original was Cohen (1988, p. 275) who calls it 'f', but all other authors label it 'r'

Since it requires a z-value the same settings as for ord_ts_wilcox function are needed.

The formula is fairly straightforward:
\deqn{r = \frac{z}{\sqrt{n}}}

The rule of thumb for the qualification is taken from Bartz (1999, p. 184):\tabular{ll}{
   r \tab qualification \cr
   0 < .20 \tab very low \cr
   0.20 < 0.40 \tab low \cr
   0.40 < 0.60 \tab moderate \cr
   0.60 < 0.80 \tab strong \cr
   0.80 or more \tab very strong \cr
}
}
\references{
Bartz, A. E. (1999). \emph{Basic statistical concepts} (4th ed). Merrill.

Cohen, J. (1988). \emph{Statistical power analysis for the behavioral sciences} (2nd ed). L. Erlbaum Associates.

Fritz, C. O., Morris, P. E., & Richler, J. J. (2012). Effect size estimates: Current use, calculations, and interpretation.
\emph{Journal of Experimental Psychology: General, 141}(1), 2–18. doi:10.1037/a0024338

Mangiafico, S. S. (2016). \emph{Summary and analysis of extension program evaluation in R (1.15.0)}. Rutger Cooperative Extension.

Tomczak, M., & Tomczak, E. (2014). The need to report effect size estimates revisited. An overview of some recommended measures of effect size.
\emph{Trends in Sport Sciences, 1}(21), 19–25.
}
