% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleOrdinal.R
\name{ord_ts_wilcox}
\alias{ord_ts_wilcox}
\title{Single Ordinal - Test - Wilcoxon
.
Performs a one-sample Wilcoxon signed rank test}
\usage{
ord_ts_wilcox(
  data,
  field,
  order = "none",
  hypMed = "none",
  exact = NULL,
  correct = FALSE
)
}
\arguments{
\item{data}{A dataframe to be used}

\item{field}{Specific field in the dataframe as string}

\item{order}{A vector with the labels in order, or set to 'none'}

\item{hypMed}{The median according to the null hypotheses. If set to none the median of the categories will be used}

\item{exact}{Boolean to use the exact Wilcoxon distribution or not, only applicable if there are no ties}

\item{correct}{Boolean to indicate the use of a Yates continuity correction}
}
\value{
results A dataframe with:

zVal The z-value statistic

pVal The two-sided p-value
}
\description{
Single Ordinal - Test - Wilcoxon
.
Performs a one-sample Wilcoxon signed rank test
}
\references{
Wilcoxon, F. (1945). Individual comparisons by ranking methods. \emph{Biometrics Bulletin, 1}(6), 80. doi:10.2307/3001968
}
