% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleOrdinal.R
\name{ord_vi_barstack}
\alias{ord_vi_barstack}
\title{Single Ordinal - Visualisation - Single Stack Bar chart}
\usage{
ord_vi_barstack(data, field, order = "none", height = "count")
}
\arguments{
\item{data}{A dataframe to be used}

\item{field}{Specific field in the dataframe as string}

\item{order}{A vector with the labels in order, or set to 'none'}

\item{height}{Either 'count' or 'percent'}
}
\description{
This function will return a stacked bar-chart of one field.
It can be defined as “a bar chart showing multiple bars stacked at each x-axis category,
each representing a value of the stacking variable” (Upton & Cook, 2014, p. 88)
}
\references{
Upton, G. J. G., & Cook, I. (2014). \emph{Dictionary of statistics} (3rd ed.). Oxford University Press.
}
